+++ CHANGELOG.md
@@ -1,15 +1,18 @@
 # Changelog
 
-## latest
+## main
+
+* Add find_user to api.jl
+* Rename examples/get-userid.jl to examples/find-user.jl
 * Added `extraHeaders` support to the SDK. This could be used as following:
 ```
 create_engine(ctx, engine; size = size, headers=["x-rai-parameter-compute-version" => "....-rai-server-binary-linux"])
 ```
 
 ## v0.0.2
+
 * Added synchronous `exec()` function that polls the v2 `exec_async()` function until completion ([#25](https://github.com/RelationalAI/rai-sdk-julia/pull/25)).
     - Uses the "v2 protocol", so the transactions will show up in your transaction log.
     - If you cancel the polling via `ctrl-C`, the error log will print the transaction ID, so you can still
       recover the transaction or cancel it.
 * Consistent return format (`Dict`) from `exec_async()` and `exec()`, regardless of whether you get synchronous results ([#24](https://github.com/RelationalAI/rai-sdk-julia/pull/24)).
-